fix: 开放题上次分数和次数

+++ internal/app/model/response/open_quest.go
@@ -28,7 +28,7 @@ type UserOpenQuestJsonElements struct {
 	UserScore             int64          `gorm:"column:user_score" form:"user_score" json:"user_score"`    // 用户分数
 	NickName              *string        `gorm:"column:nickname;type:varchar(200);default:''" json:"nickname" form:"nickname"`
 	SubmitCount           int64          `gorm:"column:submit_count" json:"submit_count"`
-	LastScore             int64          `gorm:"column:last_score" json:"last_score"`
+	LastScore             float64        `gorm:"column:last_score" json:"last_score"`
 }
 
 type UserOpenQuestJsonElementsV2 struct {
